redid internals
made everything very dynamic rather than being static.
allows for more flexibility without having to create a bunch of different static things.

+++ quasar/src/qs/utils/TextureFilters.hpp
@@ -17,7 +17,7 @@ namespace qs
 	///
 	/// Enum contaning GL texture filtering options.
 	///
-	enum class TextureFilter : int
+	enum class TextureFilter : short
 	{
 		///
 		/// Produces blocky/pixellated texture patterns.
